+++ Framework/Quickblox.framework/Versions/A/Headers/MessagesService/Classes/Business/Models/PushMessageBase/QBMPushMessageBase.h
@@ -16,6 +16,6 @@
 
 - (id)initWithPayload:(NSDictionary *)payload;
 - (NSString *)json;
-- (int)charsLeft;
+- (NSInteger)charsLeft;
 
 @end
